# [SPARK-39679][SQL] TakeOrderedAndProjectExec should respect child output ordering