--- a/sql/core/benchmarks/TakeOrderedAndProjectBenchmark-jdk11-results.txt
+++ b/sql/core/benchmarks/TakeOrderedAndProjectBenchmark-jdk11-results.txt
@@ -2,10 +2,10 @@
 TakeOrderedAndProject
 ================================================================================================
 
-OpenJDK 64-Bit Server VM 1.8.0_332-b09 on Linux 5.13.0-1031-azure
+OpenJDK 64-Bit Server VM 11.0.15+10-LTS on Linux 5.13.0-1031-azure
 Intel(R) Xeon(R) Platinum 8370C CPU @ 2.80GHz
 TakeOrderedAndProject with SMJ:           Best Time(ms)   Avg Time(ms)   Stdev(ms)    Rate(M/s)   Per Row(ns)   Relative
 ------------------------------------------------------------------------------------------------------------------------
-TakeOrderedAndProject with SMJ                      256            285          43          0.0       25609.0       1.0X
+TakeOrderedAndProject with SMJ                      330            448         104          0.0       33023.7       1.0X