--- a/sql/core/benchmarks/TakeOrderedAndProjectBenchmark-jdk17-results.txt
+++ b/sql/core/benchmarks/TakeOrderedAndProjectBenchmark-jdk17-results.txt
@@ -2,10 +2,10 @@
 TakeOrderedAndProject
 ================================================================================================
 
-OpenJDK 64-Bit Server VM 1.8.0_332-b09 on Linux 5.13.0-1031-azure
-Intel(R) Xeon(R) Platinum 8370C CPU @ 2.80GHz
+OpenJDK 64-Bit Server VM 17.0.3+7-LTS on Linux 5.13.0-1031-azure
+Intel(R) Xeon(R) Platinum 8272CL CPU @ 2.60GHz
 TakeOrderedAndProject with SMJ:           Best Time(ms)   Avg Time(ms)   Stdev(ms)    Rate(M/s)   Per Row(ns)   Relative
 ------------------------------------------------------------------------------------------------------------------------
-TakeOrderedAndProject with SMJ                      256            285          43          0.0       25609.0       1.0X
+TakeOrderedAndProject with SMJ                      384            437          47          0.0       38381.8       1.0X