--- a/sql/core/benchmarks/TakeOrderedAndProjectBenchmark-results.txt
+++ b/sql/core/benchmarks/TakeOrderedAndProjectBenchmark-results.txt
@@ -3,9 +3,10 @@ TakeOrderedAndProject
 ================================================================================================
 
 OpenJDK 64-Bit Server VM 1.8.0_332-b09 on Linux 5.13.0-1031-azure
-Intel(R) Xeon(R) Platinum 8370C CPU @ 2.80GHz
-TakeOrderedAndProject with SMJ:           Best Time(ms)   Avg Time(ms)   Stdev(ms)    Rate(M/s)   Per Row(ns)   Relative
-------------------------------------------------------------------------------------------------------------------------
-TakeOrderedAndProject with SMJ                      256            285          43          0.0       25609.0       1.0X
+Intel(R) Xeon(R) Platinum 8171M CPU @ 2.60GHz
+TakeOrderedAndProject with SMJ:                    Best Time(ms)   Avg Time(ms)   Stdev(ms)    Rate(M/s)   Per Row(ns)   Relative
+---------------------------------------------------------------------------------------------------------------------------------
+TakeOrderedAndProject with SMJ for doExecute                 287            337          43          0.0       28734.9       1.0X
+TakeOrderedAndProject with SMJ for executeCollect            150            170          30          0.1       15037.8       1.9X