--- a/sql/core/src/test/scala/org/apache/spark/sql/execution/benchmark/TakeOrderedAndProjectBenchmark.scala
+++ b/sql/core/src/test/scala/org/apache/spark/sql/execution/benchmark/TakeOrderedAndProjectBenchmark.scala
@@ -44,7 +44,7 @@ object TakeOrderedAndProjectBenchmark extends SqlBasedBenchmark {
 
     val benchmark = new Benchmark("TakeOrderedAndProject with SMJ", row, output = output)
 
-    benchmark.addCase("TakeOrderedAndProject with SMJ", 3) { _ =>
+    benchmark.addCase("TakeOrderedAndProject with SMJ for doExecute", 3) { _ =>
       withSQLConf(SQLConf.AUTO_BROADCASTJOIN_THRESHOLD.key -> "-1",
           SQLConf.SHUFFLE_PARTITIONS.key -> "5",
           SQLConf.ADAPTIVE_EXECUTION_ENABLED.key -> "false") {
@@ -52,6 +52,17 @@ object TakeOrderedAndProjectBenchmark extends SqlBasedBenchmark {
           .orderBy(col("c1"))
           .limit(100)
           .noop()
+      }
+    }
+
+    benchmark.addCase("TakeOrderedAndProject with SMJ for executeCollect", 3) { _ =>
+      withSQLConf(SQLConf.AUTO_BROADCASTJOIN_THRESHOLD.key -> "-1",
+          SQLConf.SHUFFLE_PARTITIONS.key -> "5",
+          SQLConf.ADAPTIVE_EXECUTION_ENABLED.key -> "false") {
+        df1.join(df2, col("c1") === col("c2"))
+          .orderBy(col("c1"))
+          .limit(100)
+          .collect()
       }
     }
     benchmark.run()