@ulysses-you 2022-07-05
@wangyum sure, added the test using a local sort before the TakeOrderedAndProjectExec